Converted the translation system to TypeScript.
