+++ meteor-app/imports/ui/translations/es.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import * as React from "react";
 
 export default {
 	homePage: {
