--- meteor-app/imports/ui/translations/index.js
@@ -1,11 +1,0 @@
-import makeTranslationComponents from "./translations.jsx";
-
-import en from "./en.jsx";
-import es from "./es.jsx";
-
-export default makeTranslationComponents({
-	en,
-	es,
-});
-
-export { makeTranslationComponents };
