+++ meteor-app/imports/ui/translations/index.ts
@@ -1,0 +1,11 @@
+import makeTranslationComponents from "./translations";
+
+import en from "./en";
+import es from "./es";
+
+const translations = makeTranslationComponents({
+	en,
+	es,
+});
+
+export { translations };
